+++ internal/repositories/orderrepository/orderRepository.go
@@ -10,6 +10,6 @@ type OrderRepository interface {
 	FindByNumber(ctx context.Context, number int) (entity.Order, error)
 	Save(ctx context.Context, order entity.Order) error
 	GetAllByUserID(ctx context.Context, userID string) ([]entity.Order, error)
-	GetTotalAccrualByUserID(ctx context.Context, userID string) (float64, error)
+	GetTotalAccrualByUserID(ctx context.Context, userID string) (float32, error)
 	StopSchedulerAndWorkerPool()
 }
